requirements.txt
matplotlib==3.3.1
numpy==1.18.5
opencv_python==4.2.0.32
Pillow==9.0.1
pynput==1.7.1
PyQt5==5.15.6
pyqtgraph==0.11.0
pyserial==3.5
scikit_learn==1.0.2
scipy==1.5.0
torch==1.7.1
torchvision==0.8.2
